% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{decode}
\alias{decode}
\title{Decodes an ID back into a vector of unsigned integers}
\usage{
decode(id, options)
}
\arguments{
\item{id}{Encoded ID}

\item{options}{List of settings generated by sqids_options}
}
\value{
Vector of unsigned integers
}
\description{
These are the cases where the return value might be an empty array: \cr
- Empty ID / empty string \cr
- Non-alphabet character is found within ID
}
