% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{encode}
\alias{encode}
\title{Encodes a vector of unsigned integers into an ID}
\usage{
encode(numbers, options)
}
\arguments{
\item{numbers}{Vector of non-negative numbers to encode}

\item{options}{List of settings generated by sqids_options}
}
\value{
ID string
}
\description{
These are the cases where encoding might fail: \cr
- One of the numbers passed is smaller than 0 or greater than ".Machine$integer.max" \cr
- An n-number of attempts has been made to re-generated the ID, where n is alphabet length + 1
}
