% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{encode_numbers}
\alias{encode_numbers}
\title{Internal function that encodes a vector of unsigned integers into an ID}
\usage{
encode_numbers(numbers, options, increment = 0)
}
\arguments{
\item{numbers}{Non-negative integers to encode into an ID}

\item{options}{List of settings generated by sqids_options}

\item{increment}{Internal number used to modify the 'offset' variable in order to re-generate the ID}
}
\value{
ID string
}
\description{
Internal function that encodes a vector of unsigned integers into an ID
}
