% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{is_blocked_id}
\alias{is_blocked_id}
\title{Internal function to check if generated ID includes blocked word}
\usage{
is_blocked_id(id, blocklist)
}
\arguments{
\item{id}{ID string}

\item{blocklist}{Vector of blocked strings}
}
\value{
Boolean
}
\description{
Internal function to check if generated ID includes blocked word
}
