% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{max_value}
\alias{max_value}
\title{Internal function to quickly access biggest possible integer value}
\usage{
max_value()
}
\value{
Maximum integer value
}
\description{
Internal function to quickly access biggest possible integer value
}
