% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{shuffle}
\alias{shuffle}
\title{Internal function to consistently shuffle alphabet}
\usage{
shuffle(alphabet)
}
\arguments{
\item{alphabet}{Alphabet string}
}
\value{
Shuffled alphabet string
}
\description{
Internal function to consistently shuffle alphabet
}
