% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{split_and_reverse}
\alias{split_and_reverse}
\title{Internal function to rearrange the alphabet at an offset, then reverse it.}
\usage{
split_and_reverse(alphabet, offset)
}
\arguments{
\item{alphabet}{Alphabet string}

\item{offset}{Integer offset of start of second half}
}
\description{
Before reversing the input is split based off the offset position and the
first half is put after the second.
  - First half: [1, offset - 1)
  - Second half: [offset, nchar(alphabet)]
}
