% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{sqids_defaults}
\alias{sqids_defaults}
\alias{DEFAULT_ALPHABET}
\alias{DEFAULT_MIN_LENGTH}
\alias{DEFAULT_BLOCKLIST}
\title{Default values for sqids settings}
\format{
An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 560.
}
\usage{
DEFAULT_ALPHABET

DEFAULT_MIN_LENGTH

DEFAULT_BLOCKLIST
}
\description{
Default alphabet, minimum length, and blocklist values.
}
\keyword{datasets}
