% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{sqids_options}
\alias{sqids_options}
\title{A function to create a list of sqids settings}
\usage{
sqids_options(
  alphabet = DEFAULT_ALPHABET,
  min_length = DEFAULT_MIN_LENGTH,
  blocklist = DEFAULT_BLOCKLIST
)
}
\arguments{
\item{alphabet}{String of characters for IDs}

\item{min_length}{Minimum length for IDs}

\item{blocklist}{Vector of strings to block in generated IDs}
}
\value{
A list of parameters to configure encoding and decoding
}
\description{
A function to create a list of sqids settings
}
