% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{to_id}
\alias{to_id}
\title{Internal function to convert a number to it's respective character}
\usage{
to_id(num, alphabet)
}
\arguments{
\item{num}{Integer to convert}

\item{alphabet}{Alphabet string integer is mapped to}
}
\value{
ID string
}
\description{
Internal function to convert a number to it's respective character
}
