% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqids.R
\name{to_number}
\alias{to_number}
\title{Internal function to convert a ID string to it's respective number}
\usage{
to_number(id, alphabet)
}
\arguments{
\item{id}{ID string to convert}

\item{alphabet}{Alphabet string ID was mapped from}
}
\value{
Unsigned integer value
}
\description{
Internal function to convert a ID string to it's respective number
}
